refac(#15): agent 파일 리펙토링
