+++ app/services/agents/__init__.py
@@ -1,0 +1,17 @@
+# app/services/agents/__init__.py
+from .orchestrator import orchestrator_agent, OrchestratorDeps
+from .classification import classification_agent
+from ._base import preprocessor_agent
+from .memory import load_history, save_history, load_memory, save_memory, is_memory_stale
+
+__all__ = [
+    "orchestrator_agent",
+    "OrchestratorDeps",
+    "classification_agent",
+    "preprocessor_agent",
+    "load_history",
+    "save_history",
+    "load_memory",
+    "save_memory",
+    "is_memory_stale",
+]
